FELT: Adding GitHub Actions handling for push and PR
This is adding workflow to perform builds on Linux, Windows and macOS on
push as well as pull requests events.

It will triggers creation of instances in GCP to run the builds.

Scripts are provided to also prepare the base disk image for both
Windows and Linux.

Build is done for all platforms on Linux hosts using cross-compilation.